% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backslasher.R
\name{backslasher}
\alias{backslasher}
\title{Backslasher Function}
\usage{
backslasher()
}
\description{
This function allows you to replace the backslashes of the string in your clipboard by normal slashes
}
\keyword{backslash}
