% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_results.R
\name{clean_results}
\alias{clean_results}
\title{Limpieza base}
\usage{
clean_results(x)
}
\description{
This function cleans and filter base results
}
\examples{
clean_results(base_resultado)
}
\keyword{clean_results}
