% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_ninebox.R
\name{df_ninebox}
\alias{df_ninebox}
\title{df_ninebox Function}
\usage{
df_ninebox(process_id, connection = con)
}
\arguments{
\item{process_id}{The id of a process that features desempeño y potencial xd}

\item{connection}{Name of the connection to the database}
}
\description{
Downloads and prepares the data required to run the ninebox function
}
\examples{
test <- df_ninebox(3659)
test2 <- df_ninebox(1129, base_rkmi)
test3 <- df_ninebox(2581)

}
