% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_taxonomy.R
\name{gen_taxonomy}
\alias{gen_taxonomy}
\title{gen_taxonomy Function}
\usage{
gen_taxonomy(
  df,
  birthdate_colname = "birthdate",
  output_colname = "generacion"
)
}
\arguments{
\item{df}{a data frame.}

\item{output_colname}{string with the name which will contain the classification.}

\item{birthdate}{string with the name of the df column that contains the birthdates.}
}
\description{
This function allows you quickly classify people into generations based on their birthdate
}
\examples{
gen_taxonomy(base_personas)
gen_taxonomy(base_personas, fecha_nacimiento,
generation)
}
