% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninebox.R
\name{ninebox}
\alias{ninebox}
\title{ninebox Function}
\usage{
ninebox(
  df_list,
  upper_desempeno,
  lower_desempeno,
  upper_potencial,
  lower_potencial,
  col_desempeno = "desempeño",
  col_potencial = "potencial",
  transparency = 0.2
)
}
\arguments{
\item{df_list}{Output of the df_ninbox function.}

\item{upper_desempeno}{Upper boundary for desempeño axis as dbl.}

\item{lower_desempeno}{Lower boundary for desempeño axis as dbl.}

\item{upper_potencial}{Upper boundary for potencial axis as dbl.}

\item{lower_potencial}{Lower boundary for potencial axis as dbl.}

\item{col_desempeno}{Name of the column containing desempeño as string.}

\item{col_potencial}{Name of the column containing potencial as string.}

\item{transparency}{Manipulates the alpha parameter from ggplots geom_point.}
}
\description{
Models the distribution of the ninebox with the specified boundaries
}
\examples{
ninebox(test, 90, 85, 100, 85, "desempeño ranco", "potencial ranco")
ninebox(test2, 3, 1.5, 2, 1.5)
ninebox(test3, 4, 2, 4, 2, "evaluación de desempeño", "evaluación de potencial", 0.1)
}
