% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trash_enterprises.R
\name{trash_enterprises}
\alias{trash_enterprises}
\title{trash_enterprises Function}
\usage{
trash_enterprises()
}
\description{
Returns shitty enterprises from Rankmi Database (MANUELITA QUË ES ESA WEÁ BRO)
}
