% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_taxonomy.R
\name{gen_taxonomy}
\alias{gen_taxonomy}
\title{gen_taxonomy Function}
\usage{
gen_taxonomy(
  df,
  birthdate_colname = "birthdate",
  output_colname = "generacion"
)
}
\description{
This function allows you quickly classify people into generations based on their birthdate
}
\keyword{backslash}
