% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trash_enterprises.R
\name{trash_enterprises}
\alias{trash_enterprises}
\title{trash_enterprises Function}
\usage{
trash_enterprises()
}
\description{
Returns test enterprises from Rankmi Database (MANUELITA QUÉ ES ESA WEÁ BRO)
}
