% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trash_surveys.R
\name{trash_surveys}
\alias{trash_surveys}
\title{trash_surveys Function}
\usage{
trash_surveys(df, var)
}
\arguments{
\item{df}{Data frame to filter}

\item{var}{Name of the df var to filter as string}
}
\description{
Filters test surveys from Rankmi Database
}
